REF Move image sequence tests to separate file
